pi_status/requirements.txt
luma.core==2.4.1
luma.oled==3.13.0
qrcode==7.4.2
